requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
mysql-connector-python==9.5.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.21
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
